Begin typescript migration
